added emp id

+++ src/main/java/com/sean/taller/repository/EmployeeRepository.java
@@ -2,6 +2,6 @@ package com.sean.taller.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.sean.taller.model.hr.Employee;
 
-public interface EmployeeRepository extends CrudRepository<Employee, Long>{
+public interface EmployeeRepository extends CrudRepository<Employee, Integer>{
 
 }
